refactor: UserWithActivityUnit을 domain/model로 이동
