+++ src/test/kotlin/co/yappuworld/support/fixture/AttendanceFixture.kt
@@ -13,7 +13,7 @@ import co.yappuworld.support.fixture.UserFixture.getUserWithActivityUnitsFixture
 import co.yappuworld.user.domain.model.UserWithActivityUnits
 import co.yappuworld.user.domain.vo.Position
 import co.yappuworld.user.domain.vo.UserRole
-import co.yappuworld.user.infrastructure.model.UserWithActivityUnit
+import co.yappuworld.user.domain.model.UserWithActivityUnit
 import java.time.LocalDateTime
 import java.util.UUID
